Add Account model, stop modifying autoload_paths
The developer will almost certainly want to retrieve user accounts with
ActiveRecord, and I think they could be thrown off if there are no
ActiveRecord models generated, so we generate the Account model for them.

We also leave the responsibility of adding lib/ to autoload_paths to the
developer, because I think it's too magical if the generator edits
config/application.rb.