Use #change migration instead of #up
Previously we had to use #up and #down because we had `#execute`.
However, now we should be able to use #change. Hopefully ActiveRecord
reverses the order in which it drops tables, or applies cascading,
otherwise the down migration won't work. We'll test that out shortly,
and revert the change if it doesn't work.