Add rodauth:mailer generator for generating mailer with views
This makes it easier to get started with overriding mailer views. I
chose to make it separate from rodauth:views generator in order to
separate these two steps mentally.

The user is still required to modify the Rodauth app configuration to
call the new mailer, but they already have the code ready that they just
need to comment out.

We might add the HTML versions in the future, but for now we have only
text versions because those are the ones that ship with Rodauth.