Use :action instead of :template when rendering
Using :template requires us to specify the full path, and I'm not sure
how to retrieve the view path for a given controller. It seems that
`render "template"` internally resolves to `render action: "template"`,
so we use that, that way we're still able to set the :layout option.

It's a bit weird that we're using :action even though we're rendering
outside of a controller action, but hey, that's Rails.