Allow Rails app to call Rodauth methods that halt
When a Rodauth method redirects, internally this is implemented in Roda
by throwing :halt. Roda's middleware doesn't currently catch :halt
around the downstream app, so if we tried to call a Rodauth method that
redirects from within a Rails app, we would get uncaught :halt error.

So we catch :halt in our middleware. The throw value will be the Rack
response triple, so we can just let that be the return value.