Commit Rails flash before redirect
This allows Rails app to call Rodauth methods that do redirects (such as
`#require_login` and `#require_authentication`), while still capturing
redirect flash messages.

When a Rodauth redirect is called within the Rails application, the
:halt throw goes outside of the Roda app (and is caught in
Rodauth::Rails::Middleware), so the after block would never get called.
Therefore we make sure to commit the Rails flash before redirect.

I was kind of hoping that the flash middleware in Rails already takes
care of this, but I think it only calls #commit_flash on the controller
level, so that doesn't get called either.