Make Sequel reuse ActiveRecord's connection
Previously we were creating a separate connection with Sequel, because I
was under the impression it was impossible for Sequel to reuse
ActiveRecord's connection.

However, it turns out it's not. We achieve it with the following steps:

* use fake connection pool that just returns ActiveRecord's connection
* wrap ActiveRecord connection in object that conforms to Sequel API
  - #execute optionally yields results
  - convert ActiveRecord::NotFound into Sequel equivalents
* use ActiveRecord's transaction implementation
  - convert Sequel::Rollback into ActiveRecord::Rollback
* avoid testing the connection to avoid establishing AR connection

This is a great achievement. This means the caveats related to
performance and transactions are now gone.

People might still complain about having to load Sequel, but since
Sequel is 5x lighter than ActiveRecord, in reality the impact should be
minimal.