Finish tests and the initial implementation
We finish tests, both for generators and for integration with a Rails
app.

We rename `.rodauth` class method to `.configure` to avoid clashing with
Rodauth plugin's `.rodauth` method.

We add support for Rodauth 1.23, for folks who cannot upgrade to 2.0
yet. The reason we cannot support earlier versions is because 1.23 added
`#send_email` which our Rails feature is using.

We also do many other smaller fixes that wraps up the implementation.