Generate RodauthController on install, remove RodauthHelper
Generating RodauthController on install means the user doesn't have to
manually edit Rodauth app configuration when they're generating views,
because we can already do this on install. It also means the user
doesn't have to see it in the output if they run this command later on
for more features.

We also remove RodauthHelper, as the #locals method is really not
needed.

We remove the `-f` abbreviation for `--features` because it clashes with
the same abbreviation for `--force` (causing `-f` to be interpreted as
`--force` instead of `--features`).