Use #password_field_autocomplete_value in password_field partial
Rodauth 2.1 will add a new #password_field_autocomplete_value method
which returns the appropriate autocomplete value for the password field
depending on the action. For create account and verify account actions
it will return "new-password", otherwise it will return
"current-password". So, we update the password_field partial to use this
method on Rodauth >= 2.1.

See the commit and pull request for more details:

* https://github.com/jeremyevans/rodauth/commit/2b6ce2539f3fc3f2d7dcb45a2dbd46ecea66fddb
* https://github.com/jeremyevans/rodauth/pull/98