Suggest sending emails in the background after transaction commit
So far we were suggesting delivering emails synchronously, which is
obviously brittle, so we recommend delivering in the background instead.
Now that sequel-activerecord_connection gem supports transaction hooks,
we can use them to queue the jobs after the transaction commits, so that
the background job has the latest changes in case of any queries.