Recommend a more development-friendly #current_account setup
In development accounts often get deleted for testing purposes, which
causes annoying errors if the code doesn't handle the fact that the
account ID in the session doesn't have to exist.

So we recommend loading account for authenticated requests, but handling
the case when account was not found, by logging the session out and
requiring login. This is how the Rodauth::Info#require_account private
method works.

We prefer handling this logic upfront, rather than lazily when the
`#current_account` method is called.