Support Rails API-only mode
When API-only mode is enabled in Rails, the ActionDispatch::Flash
middleware doesn't get added, and the ActionDispatch::Request object
doesn't define flash methods. This was problematic, because our flash
Roda plugin try to call these methods in before/after hooks.

Instead of modifying the flash plugin to handle that, we skip loading it
when `json: :only` is specified on configuration. We also pass `json:
:only` in rodauth_app.rb when Rails API-only is configured, as well as
add jwt feature and remove remember, and set JWT secret.