Bump JWT instructions up and improve them
JSON API support is pretty important, and it's one of the main
advantages of Rodauth, so it should be much higher than below the "How
it works" heading.

Also show adding the JWT gem to the Gemfile, and clarify the difference
between `json: true` and `json: :only` just in case.