Don't use RodauthController in API-only mode
The RodauthController is used for rendering and CSRF protection, neither
of which are used in API-only mode. Since the intention of the API-only
mode is to provide a more barebones tools, let's respect that and
skip RodauthController in this case.

In the Rodauth extension, we change rendering to be skipped in case of
`json: :only` mode. Rodauth will still use the #render method for email
body creation, and we don't want this to be trying to call
ActionController::Base descendants.

I was tempted to disable CSRF-protection as well, but the JWT feature
already turns off #check_csrf?, so there should be no need.