Set default bcrypt hash cost to 1 in tests
Rodauth reads RACK_ENV environment variable to set default bcrypt hash
cost, which is 1 when the value is "test". It appears Rails doesn't set
this variable in tests (only RAILS_ENV is set in test_helper.rb), so we
have to set it ourselves.

This speeds up the test suite 4 times.