Drop support for Ruby 2.2
We've dropped it in sequel-activerecord_connection 1.1.0, and since we
want to depend on that version or higher due to significant
improvements, we drop support for Ruby 2.2 in rodauth-rails as well.

I don't remember why we went all the way to supporting Ruby 2.2. I
thought that was the maximum version on which Rails 4.2 worked, but
we *are* successfully testing Rails 4.2 support on Ruby < 2.7 already,
so that wasn't it.