Add rodauth:routes rake task for printing routes
One of the pain points when getting started with Rodauth is not knowing
what routes it handles. This is because Rodauth uses Roda for routing,
so even if we were able to hook into the `rails routes` command, we
wouldn't be able to programmatically retrieve the routes.

So, we add a new "rodauth:routes" rake task which prints all routes
handled by the current Rodauth configuration.

  $ rails rodauth:routes
  Routes handled by Rodauth::Rails::Middleware:

    GET/POST  /login                   rodauth.login_path
    GET/POST  /create-account          rodauth.create_account_path
    GET/POST  /verify-account-resend   rodauth.verify_account_resend_path
    GET/POST  /verify-account          rodauth.verify_account_path
    GET/POST  /change-password         rodauth.change_password_path
    GET/POST  /change-login            rodauth.change_login_path
    GET/POST  /logout                  rodauth.logout_path
    GET/POST  /remember                rodauth.remember_path
    GET/POST  /reset-password-request  rodauth.reset_password_request_path
    GET/POST  /reset-password          rodauth.reset_password_path
    GET/POST  /verify-login-change     rodauth.verify_login_change_path
    GET/POST  /close-account           rodauth.close_account_path

The 1st column are possible HTTP verbs, the 2nd column is the route, and
the 3rd column is the method for retrieving the route path. It handles
several things:

* displays route prefix if configured
* displays correct code for secondary configurations
* excludes GET routes on API-only mode

I chose not to use `Rodauth.route_hash`, because it doesn't retrieve the
HTTP verbs nor information on route names.