Make rodauth:routes work with external features
Manually handling each feature won't work with external features. So we
read from the Rodauth.routes.

The user loses the HTTP verbs for built-in routes, but that's the
trade-off for consistency and reduced complexity. And that was never
runtime information to begin with. But I don't think the HTTP verbs were
important anyway.

We also write tests for the rodauth:routes rake task.