Add rodauth:migration generator
Previously, the rodauth:install generator would include create_table
statements for all features, where the statements for non-default
features were commented out. This meant that people that deleted the
comments would have to create new Rodauth tables, which meant they'd
either have to translate Sequel code, or look into rodauth-rails source
code.

To remedy that, we add a new rodauth:migration generator which generates
a migration with create_table statements for selected features.

  $ rails generate rodauth:migration otp sms_codes recovery_codes

For this we extract the migration-related code out of the install
generator, in order to be able to reuse it in the migration generator.
We also extract create_table statements for individual features into
separate *.erb files.