Add example of setting HMAC secret
People should probably set the HMAC secret to something different than
Rails' secret key base, as that allows it to be rotated separately when
needed. So we show an example of that, to raise awareness that this setting
exists, while keeping Rails' secret key base as the default. This is
what Devise does as well.