Support Active Record adapters in rodauth:install
Since most external Active Record adapters have an equivalent Sequel
adapter, we should always generate the sequel.rb initializer, even if we
cannot guarantee the Sequel adapter is called the same. We add mappings
for Oracle and SQL Server, which should take care of most cases. We use
Sequel.connect so that the user gets an explicit AdapterNotFound error
if the adapter name doesn't match.