Support Rack < 2 in Rodauth::Rails.rodauth
It appears that Rack < 2 requires the port to be set, otherwise it will
default it to "0", generating URLs like https://example.com:0/login. So
we assign the default port for the configured URL scheme.

Rails 4.2 uses Rack < 2, and possibly some Rails applications haven't
yet upgraded to Rack 2.x due to backwards incompatibilities, so we
should probably support it even if we drop support for Rails 4.2.