Fix remember cookie deadline not extending in remember feature
If the session cookie has expired (e.g. the browser was closed) and the
remember cookie is still present, the `rodauth.load_memory` call will
autologin the user from the remember cookie. If the
`extend_remember_deadline?` setting is true, the remember cookie
expiration will be extended as well.

Extending the remember cookie is done by setting the `Set-Cookie`
response header. The problem is, if Rodauth doesn't handle the request
that remembers the user from the token (which is the most likely
scenario), the request is forwarded to the Rails app, and any response
headers that have been set in the Rodauth app's block is thrown away,
including `Set-Cookie`.

We work around this by saving the response headers to the env hash if
Roda will not handle the current request, and then after the Rails app
has returned the response, we merge any headers set in the route block
into the final response.

This should not override any cookies potentially set by the Rails
application, because the ActionDispatch::Cookie middleware is the one
that adds the `Set-Cookie` header after a request has finished. This
happens after we've set our `Set-Cookie` header, and the middleware is
smart to append to any existing header value.