Move JSON API instructions to a separate section
rodauth-rails technically doesn't provide any integration here, other
than generating this when the Rails app is using API-only mode. I find
it better that it's in a separate section, as the "Usage" section is
already quite long now.

Also show `json: :only` in the example, as I think that will be the most
common case.