Suggest checking for #logged_in?
If the account was deleted manually, it can happen that with otp,
email_auth, and verify_account_grace_period feature the #authenticated?
call fails if the account id is still written to the session. Therefore
it's safer to check for #logged_in?, we don't need to check for MFA
anyway.