Don't blow up on unsupported HTTP verbs
Rodauth calls the `#_around_action` block inside a `request.is` block,
and if the HTTP verb is unsupported, the handle method will simply
return nil. This caused rodauth-rails to blow up, as it expected result
to be a Rack triple. We handle that by returning nil and letting
Rodauth's router return 404.