Reset Rails session on #clear_session
This is standard practice that both Devise and Sorcery employs, and it
apparently protects from session fixation attacks. I'm still not able to
fully understand this attack, and I think that way that Rails works now,
with session ID being part of the session payload, prevents these kind
of attacks.

But just to be certain, we call `#reset_session` to reset the session
ID. We also require Rodauth 2.8+, as it includes a fix for JWT that
makes clearing session data work with our new `#clear_session` override.