Add Rodauth::Rails::Auth superclass
This allows users to move configurations into subclasses of
Rodauth::Rails::Auth, which can then live in separate files.

  class RodauthMain < Rodauth::Rails::App
    configure do
      # ...
    end
  end

  class RodauthApp < Rodauth::Rails::App
    configure RodauthMain

    route do |r|
      # ...
    end
  end

It uses the new :auth_class plugin option, and the ability to clone the
configuration object. The Rodauth::Rails::Auth class also extends
Rodauth::Auth with the ability to do multi-level inheritance.

This will likely become the default way in rodauth-rails 1.0.