Add Rodauth::Rails.authenticated routing constraint
The documented routing constraint didn't work, because the block needs
to return a truthy value in order to fulfill the constraint, and
`#require_authentication` returns nil when the user is authenticated.

To make adding an authenticated routing constraint easier, we create a
`Rodauth::Rails.authenticated` helper method that does the logic for us.
We allow specifying the configuration and an additional condition.

Maybe we will add a .logged_in constraint as well in the future, but I
didn't think it would be very common at the moment.

Closes #32