Merge rodauth:mailer into rodauth:install
There is no good reason why mailer generation should happen separately
from the rodauth:install generator, because the user will always want to
override default email templates, and we want to encourage them to do
so. Using an Action Mailer class is a standard Rails way to do so.

So, we create the mailer and templates in rodauth:install generator,
and remove the rodauth:mailer generator. If the user is using a
3rd-party service for sending emails, they will just replace the mailer
with their code. In any case, that's easier for them to do when they see
the default email templates.