Improve Roda middleware inspect output
Roda's middleware plugin internally creates an anonymous subclass of the
Roda class, and uses it to call the app. This is then shown in the inspect
output when viewing the Rodauth instance, without any visible connection
to the main Roda class (`RodauthApp` by default).

In our setup, the anonymous Roda subclass is created for each request,
because we have Rodauth::Rails::Middleware in between RodauthApp and
Rails' middleware stack. This means we cannot assign it to a constant,
because it will be different for each call.

So instead we override #inspect to show the connection with RodauthApp.
Once we're able to create named Rodauth::Auth subclasses, there will be
no anonymous classes in the Rodauth instance inspect output anymore.