Load pass plugin and recommend calling `r.pass` for prefixed routes
By default, if `r.on("admin") { r.rodauth(:admin) }` is used, and a
request is made to an `/admin/*` endpoint which is not one of Rodauth's
endpoint, Roda will return a 404 instead of forwarding the request to
the Rails app.

Calling `r.pass` at the end of the `r.on` block tells Roda to keep
routing, which will eventually get it to pass the request to the Rails
app. For convenience we automatically load the pass plugin.