Add controller-like request logging (#42)
The Rails::Rack::Logger middleware already logs each request on the URL
level, and view rendering is being logged as well. However, requests to
controller actions also display:

  Processing by MyController#some_action as HTML
  # ...
  Completed 200 OK in 12ms (...more info...)

We want the same kind of nice logging for Rodauth requests, to
communicate to the developer that the RodauthApp middleware has indeed
caught the request, what was response status, and how long the whole
request took.

We implement this using the same approach that Action Pack uses to
implement controller-level logging.