Add Rodauth::Rails::Model for Active Record integration
Currently, creating an account with a password programmatically is a bit
cumbersome, mostly due to bcrypt-hashing, especially when the password
is stored in a separate table. This adds friction in development and in
tests, and also for anyone wanting to implement simple user creation by
admins. Devise, Sorcery, and has_secure_password all provide this level
of convenience.

Additionally, there are cases where it's useful to be able to directly
query associated tables (e.g. retrieving all accounts that have MFA
enabled), and it might not be obvious how to declare associations, e.g:

  class Account < ApplicationRecord
    has_one :remember_key, foreign_key: :id
  end

  class Account::RememberKey < ApplicationRecord
  end

To address this friction, we define a Rodauth::Rails::Model mixin that
can be be included into the account model:

  class Account < ApplicationRecord
    include Rodauth::Rails.model
  end

It provides password attribute and validations:

  account = Account.create(email: "user@example.com", password: "foo")
  account.password_hash #=> "$2a$12$k/Ub1I2iomi84Rac..."
  account.valid? #=> false
  account.errors[:password] #=> ["invalid password, does not meet requirements (minimum 6 characters)"]

It also defines associations based on the Rodauth configuration:

  account.remember_key #=> #<Account::RememberKey>
  account.active_session_keys #=> [#<Account::ActiveSessionKey>, ...]

Alternative configurations are supported as well:

  class Admin < ApplicationRecord
    include Rodauth::Rails.model(:admin)
  end

  Admin::RememberKey # can reference `admin_remember_keys`