Speed up rendering by limiting template formats
`ActionView::LookupContext` will by default look up templates in all
known formats (33 of them). This results in quite a few system calls,
which can slow down rendering, especially when filesystem access is
slower (e.g. when macOS with Docker).

Before processing an action, Action Controller reduces the list of
formats only to formats that the current request accepts. However, in
our case we're rendering templates outside of action processing, so this
code is not being executed. We fix this by calling it manually.