Allow generating views for specified rodauth configurations
We improve the `rodauth:views` generator by adding a `--name` option for
specifying a rodauth configuration:

  $ rails generate rodauth:views --name admin

It now automatically infers the correct template path from the
configured controller, so we remove the `--directory` option.

Since we're calling the `#rails_controller` method, we make it public.
We also explicitly fail in `Rodauth::Rails.app` if undefined rodauth
configuration is passed in.