Recommend storing account types in separate table
The approach with a boolean column works only when there is one
secondary configuration, so it won't work in general. The approach of
using separate tables is also not ideal, because the tables between
account types are duplicated for the same features.

So we instead recommend storing account types in a separate table, which
is inline with the Rodauth style. Like most recommendations in the
README, I haven't actually been able to test this, but people who try it
out will let me know of any issues.