Add Rodauth::Rails::Model for Active Record integration (#45)
This adds a module builder that defines a password attribute and
associations when included into an account Active Record model.

  class Account < ActiveRecord::Base
    include Rodauth::Rails.model
  end

The password attribute automatically hashes the password, and handles
both the password hash being in a separate column or a separate table.

  account = Account.new(email: "user@example.com", password: "secret")

  # when password hash is stored in a column on the accounts table
  account.password_hash #=> "$2a$12$k/Ub1I2iomi84RacqY89Hu4.M0vK7klRnRtzorDyvOkVI.hKhkNw."

  # when password hash is stored in a separate table
  account.password_hash #=> #<Account::PasswordHash...> (record from `account_password_hashes` table)
  account.password_hash.password_hash #=> "$2a$12$k/Ub1..." (inaccessible when using database authentication functions)

The associations are defined according to the enabled authentication
features, and the corresponding model classes are automatically defined
inside the namespace of the account model.

  # queries the `account_remember_keys` table
  account.remember_key #=> #<Account::RememberKey>

  # queries the `account_authentication_audit_logs` table
  Account::AuthenticationAuditLog.where(message: "login").group(:account_id)