Add `#rodauth_response` controller method for converting rodauth responses into controller responses
If the user has set up controller callbacks that do things after certain
actions, they might want to prevent rodauth methods called inside
controller actions from throwing response all the way up to the Rodauth
middleware, because that skips any `after_action` callbacks.

So we add a `#rodauth_response` controller method that can be used to
wrap rodauth methods that throw a response, which will catch `:halt` and
write response data into the controller response. For now we only show
it in the OmniAuth example, and in the future we can document it better,
or replace it with something better.

Closes #40