Add `rails_account_model` configuration method
In some cases the account model class cannot be inferred from the table
name, for instance when it's namespaced under module
(e.g. `Authentication::Account`). For these cases, we provide the
`rails_account_model` configuration method that allows the user to
specify the account model class manually.

While here, we also handle schemas in the accounts table name, which can
be specified with `Sequel[:auth][:accounts]`, and which returns a
`Sequel::SQL::QualifiedIdentifier` object.