Set `delete_account_on_close?` to `true` in generated rodauth_app.rb
This allows the email of a closed account to be used again, which is
especially convenient when testing on databases that don't support
partial unique indices, such as MySQL or SQLite. I often close accounts
via the UI, and it's annoying that I'm not able to use the same email
again, so I would personally benefit from this convenience.

The fact that Devise deletes the account record when destroying the
registration gives me confidence that this is an OK default.