Improve documentation for calling rodauth outside of a request
Because there was only a code example for calling Rodauth's instance
methods, for the reader that skipped paragraphs it might have seemed
that this is how internal_request feature is primarily used.

So we update this section to show how internal_request API is used, as
well as path_class_methods for generating authentication URLs, and move
the more advanced usage of calling instance methods into a subsection.

Closes #61