Improve rodauth:install defaults for JSON & JWT
When the user ran `rodauth:install --json` in an API-only app, they
would currently get JWT configured instead. This is because we
considered JWT automatically selected for API-only apps, even if the
`--json` option as passed.

It's common for people to have API-only apps with cookie sessions
enabled. In fact, that's the primary use case for Rodauth's JSON
feature.

So, we allow specifying the `--json` option for API-only apps. We also
default to JWT for API-only apps where session store is not enabled, and
default to JSON for API-only apps where session store is enabled. That
should hopefully be a reasonable default for most apps.

Concerns #66