Disable Turbo in recovery codes form template
Viewing recovery codes currently isn't Turbo compatible, because it
requires submitting a POST form that returns a 200 response, which isn't
supported by Turbo (Turbo requires form submissinos to either redirect
or render with 4xx or 5xx response).

So, we disable Turbo for the recovery codes form in a partial created by
`rodauth:install`. It's unfortunately not easy to replicate this for
Rodauth's built-in templates, so people will have to live with that, at
least for the time being.

Concerns #67