Use Rodauth methods for retrieving texts in views
This allows for full internationalization of generated views just by
loading translations for a different locale, which makes for a much
better starting point.

I've initially chosen to hardcode texts to match Devise templates, but
we also have devise-i18n which overrides those templates with
translations. So, when in doubt, better to just do what Rodauth does by
default.

We also switch to using `<%== ... %>` for outputting HTML instead of
`#html_safe`, as it looks cleaner. This should hopefully be supported by
all Rails versions, not just those that use Erubi.