Fix error when rendering built-in templates on Turbo requests
Turbo requests communicate in the `Accept` header that they support
`:turbo_stream` and `:html` formats. Normally when rendering, Rails
searches for templates in all request formats. However, when rendering
with the `:html` option, only the first format is picked, which in this
case is `:turbo_stream`. This causes the controller to search for the
layout in `turbo_stream` format, which it cannot find.

We work around this issue by explicitly passing the `:html` format to
the render call.

Closes #68