Pass only account ID and key value to mailer
When frontend is separated from the backend, these two apps are commonly
on different hostnames (e.g. `app.foo.com` and `api.foo.com`). This
doesn't play well with the current default configuration, where the
backend URL is being passed to the mailer, instead of the URL specified
in `config.action_mailer.default_url_options`.

We fix this by passing only the key value into the mailer, and letting
the mailer generate links via path class methods, which uses URL options
configured in `config.action_mailer.default_url_options`. This means
email delivery won't work anymore if this setting is not configured, so
we should probably mention it a post-install message.

By passing the raw key value, and not the HMAC'd email token, we prevent
an attacker that has gained access to the background job database (e.g.
Redis) from using these tokens to gain access to other accounts.

We also pass the account ID instead of email, so that we can retrieve
the account record easily, which will likely be used in email bodies to
display additional information about the user.

Closes #63