Disable Turbo in all built-in and generated views
Rodauth is not guaranteed to work with Turbo, some POST endpoints return
200 HTML responses instead of redirect, such as multi-phase login and
adding recovery codes. We have already disabled those in generated
templates, but third-party plugins can offer their own actions, or
override existing Rodauth actions, in a way that's not compatible with
Turbo.

So, in order to guarantee that Rodauth actions will always work, it's
safer to disable Turbo by default. Users can always enable it if they
want, if they know that an action is Turbo-compatible.