Generate rodauth_app.rb in app/misc directory
Initial versions of this gem generated the Rodauth app inside the `lib`
directory. However, that required the user to add the `lib` directory to
Zeitwerk's autoload paths, which added friction to onboarding.

The `app/lib` directory was created as an alternative to the `lib`
directory. However, the `lib` directory is intended for code that's not
strictly app-related (and could be used somewhere else), and Rodauth
code is in fact app-related (it just doesn't have a drawer). This means
keeping "lib" naming doesn't really make sense.

I was trying to find where do Rails developers usually put code that
doesn't fit in any of Rails' drawers. In this search, I encountered an
article by @jonleighton, where he mentions putting such code in
`app/misc`. That made perfect sense to me, so we're adopting this here.