Add `Rodauth::Rails::Auth.instance`
Now that we can rely on auth classes being assigned to constants, it
makes sense to move `Rodauth::Rails.rodauth` into the auth class. So,
we define an `Rodauth::Rails::Auth.instance` method, which makes it
explicit that we're creating an instance (previously we had both
`RodauthApp.rodauth` and `Rodauth::Rails.rodauth`, and from naming it
wasn't clear that the latter returns an instance).

We deprecate `Rodauth::Rails.rodauth` in favour of
`Rodauth::Rails::Auth.instance`. If users want to fetch the auth class
by the configuration name, they can still do so via
`RodauthApp.rodauth(name).instance`.