Avoid method re-definition warnings with named auth classes
When using features that define class methods in `post_configure` (such
as internal_request or path_class_methods), registering a named auth
class would trigger method re-definition warnings in verbose mode,
caused by `post_configure` being called twice on the instance of the
same class.

We fix this by passing a block to the rodauth plugin only when one was
given, which Rodauth supports for quite a while now. The block will not
be provided when registering an already configured auth class, in which
case `Rodauth::Auth.configure` will not be called (and `post_configure`
will be skipped as a result).