Revert adding `Rodauth::Rails::Auth.instance`
I'm not sure that this is the right path, as my rule of thumb was always
to avoid adding new methods to the `Rodauth::Rails::Auth` object that
don't communicate that they're specific to rodauth-rails. Now that we've
decided to reference auth classes through the Rodauth app again, I
don't think it makes much sense to add an `.instance` class method.

This reverts commit ea929b0354a2304fc06267114f0c3670b473709e.