Automatically route path prefix in `r.rodauth`
When a routing prefix is configured in Rodauth, the developer currently
also needs to remember to route the prefix before calling `r.rodauth`.

  class RodauthApp < Rodauth::Rails::App
    configure do
      path "/user"
    end

    route do |r|
      r.on "user" do
        r.rodauth
      end
    end
  end

This makes sense when using Rodauth in a project that uses Roda as the
main web framework, as then routing Rodauth requests is likely done
on the same level as routing other requests.

In Rails, however, the developer is not expected to use any additional
Roda matchers, so the `r.on` is just an extra step they need to remember
to do when adding the path prefix. Also, because the Roda app is used as
a middleware, the developer also needs to remember to exit the `r.on`
matcher to be able to handle other routes starting with that prefix.

So, for the sake of convenience, we enhance `r.rodauth` to automatically
route the path prefix if one has been set. This removes the extra step
when setting a path prefix.