Preserve flash message through consecutive redirects
The default flash behaviour both for Roda and Rails is that flash
messages that are unread are retained on the subsequent request. This
means that if a flash message is set on request A, which redirects to
request B, and request B redirects to request C, the request C would
have access to the original flash message.

We broke that default behaviour by reading the flash before each
request, both for Rodauth routes Rails application routes. This means
that a double-redirect will not preserve the original flash message,
which is the case when requiring multifactor authentication, where only
one additional factor is setup.

I don't recall what was the original reason for reading flash messages
before each request, but with it removed I couldn't produce any
unexpected behaviour (e.g. flash messages lingering longer than they
should), and none of the tests fail. So, I'm assuming it's not necessary
after all.