Ensure requiring login follows login route changes
By default, Rodauth hardcodes `require_login_redirect` to `{prefix}/login`,
which will become outdated if the login route changes. Rodauth defines
this configuration in the base feature, which means it cannot point to
`login_path`, because the login feature might not be enabled.

However, for rodauth-rails it definitely makes sense to have this in the
default configuration, as we want to maximize convenience and minimize
surprises (and most apps will have login feature enabled).