Ensure response status is logged when halting inside controller
When calling a rodauth method that throws inside a controller (e.g.
redirecting), Rails skips code that adds response status to the event
payload that's later used for logging, resulting in a blank status in
"Completed ..." message.

It does however call `#append_info_to_payload`, so we use that
opportunity to add the response status to the payload.