Move association definitions to a Rodauth method
This allows other Rodauth features to easily extend the list with their
own associations:

  module Rodauth
    Feature.define(:foo, :Foo) do
      def associations
        list = super
        list << { name: :foo, type: :one, table: foo_table, foreign_key: foo_id_column }
        list
      end
    end
  end

We also rename the association types from :has_one/:has_many to
:one/:many, to decouple it from Active Record, paving way for a Sequel
model implementation.

This design also allows the list of associations to be used by 3rd party
code more easily. And the fact that it's defined on the instance level
makes it more succinct.