Skip calling Rodauth app on asset requests
When using Sprockets or Propshaft, request to assets got to the Rails
router, and thus call any rack middlewares on the way, which includes
the Rodauth middleware. This means if someone calls
`rodauth.require_authentication` for all except certain routes, it's
likely to get called on asset requests too, which will result in
redirect responses instead of actual file contents.

We want to skip Rodauth entirely when serving assets. So, we modify the
Rodauth middleware to skip calling the Rodauth app for asset requests.