Replace model mixin with rodauth-model gem
Because the model mixin behaviour wasn't Rails-specific, but could be
used in any web frameworks that use Active Record or Sequel models, and
because the implementation and maintenance was sufficiently complex, I
decided to extract it into a separate gem.

The rodauth-model gem defines a better API for registering custom
associations for 3rd-party Rodauth features.