Avoid passing email addresses in job arguments
This should be more GDPR-safe, because it avoids having to erase log
lines for users requesting to have their personal data removed. There is
also no need to send something in the job arguments that can be easily
calculated.

An argument could be made that retrieving current data will cause the
mailer job to perform differently if data happens to change between time
of enqueuing and time of executing. However, this was already the case
when reading the recipient email from the account record, as the email
address could have changed. And this edge case is not really
problematic in my opinion.