Make Sequel base migration reversible
When reverting the migration, the change block is being executed against
a BasicObject subclass instead of a database. This class doesn't define
`#run`, as that's not reversible programmatically, so it raises a
`NoMethodError`. We ignore that error, making reverting pass.

In the future, when we add support for authentication database functions
to migration generator, we might define an explicit `down` block instead.

Closes #118